// src/generated/models/Afb_questionnairesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Afb_questionnairesafb_statutdepublication = {
  0: 'Publi_',
  1: 'Brouillon',
  747010001: 'Archiv_'
} as const;
export type Afb_questionnairesafb_statutdepublication = keyof typeof Afb_questionnairesafb_statutdepublication;
export const Afb_questionnairesafb_typededocument = {
  0: 'AML',
  1: 'SLA',
  2: 'KYC',
  3: 'EXT',
  4: 'RISK',
  747010001: 'OPS',
  747010002: 'LIBRE'
} as const;
export type Afb_questionnairesafb_typededocument = keyof typeof Afb_questionnairesafb_typededocument;
export const Afb_questionnairesstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Afb_questionnairesstatecode = keyof typeof Afb_questionnairesstatecode;
export const Afb_questionnairesstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Afb_questionnairesstatuscode = keyof typeof Afb_questionnairesstatuscode;

export interface Afb_questionnairesBase {
  "afb_auteur@odata.bind": string;
  afb_codedudocument: string;
  afb_datedepublication: string;
  afb_descriptionducontenu?: string;
  afb_questionnaireid: string;
  afb_statutdepublication: Afb_questionnairesafb_statutdepublication;
  afb_titreenanglais: string;
  afb_titreenfrancais: string;
  afb_typededocument: Afb_questionnairesafb_typededocument;
  afb_version: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Afb_questionnairesstatecode;
  statuscode?: Afb_questionnairesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Afb_questionnaires extends Afb_questionnairesBase {
  afb_auteurname?: string;
  afb_statutdepublicationname?: string;
  afb_typededocumentname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  afb_auteur?: object;
  _afb_auteur_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}